# views/swagger.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "This is my app"
host: myapp0604.herokuapp.com
basePath: /
schemes:
- "http"

tags:
- name: "API"
  description: ""

paths:
  /listitem:
    get:
      tags:
      - "API"
      summary: "Get list of all item"
      description: ""
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/listitem"
    delete:
      tags:
      - "API"
      summary: "Delete list of item"
      description: ""
      responses:
        200:
          description: "OK"
    
    post:
      tags:
      - "API"
      summary: "Create new item"
      description: ""
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "item"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/item"
      responses:
        200:
          description: "Created"
          schema:
            $ref: "#/definitions/item"
    put:
      tags:
      - "API"
      summary: "Update item"
      description: ""
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "item"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/item"
      responses:
        200:
          description: "Updated"
          schema:
            $ref: "#/definitions/item"
            
        
  /listitem/{nameitem}:
    delete:
      tags:
      - "API"
      summary: "Delete item by nameitem"
      description: ""
      operationId: ""
      produces:
      - "application/json"
      parameters:
      - name: "nameitem"
        in: "path"
        description: "name of the item that needs to be deleted"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          

definitions:
  item:
    type: "object"
    properties:
      nameitem:
        type: "string"
      value:
        type: "integer"
        format: "int64"
    xml:
      name: "item"
      
  listitem:
    type: "object"
    properties:
      item:
        $ref: "#/definitions/item"

  